pyproject.toml
[tool.poetry]
name = "drf-lms"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "4.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.2"
pillow = "^10.4.0"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.6.1"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.4.0"
stripe = "^10.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

